Theme fixes

+++ submodules/TelegramUI/TelegramUI/OpenChatMessage.swift
@@ -514,9 +514,9 @@ func openChatWallpaper(context: AccountContext, message: Message, present: @esca
                         case let .slug(slug, options, firstColor, secondColor, intensity, rotation):
                             source = .slug(slug, content.file, options, firstColor, secondColor, intensity, rotation, message)
                         case let .color(color):
-                            source = .wallpaper(.color(Int32(color.argb)), nil, nil, nil, nil, nil, message)
+                            source = .wallpaper(.color(color.argb), nil, nil, nil, nil, nil, message)
                         case let .gradient(topColor, bottomColor, rotation):
-                            source = .wallpaper(.gradient(Int32(topColor.argb), Int32(bottomColor.argb), WallpaperSettings(rotation: rotation)), nil, nil, nil, nil, rotation, message)
+                            source = .wallpaper(.gradient(topColor.argb, bottomColor.argb, WallpaperSettings(rotation: rotation)), nil, nil, nil, nil, rotation, message)
                     }
                     
                     let controller = WallpaperGalleryController(context: context, source: source)
